Inline `dbl_classify()` because it massively improves double performance
